refactor(frontend): narrow fetchedAts to Uint16Array
Change the idx fetchedAts column from Uint32Array back to Uint16Array in
idx.ts, with the matching parser-test update.

+++ frontend/src/js/idx.ts
@@ -3,7 +3,7 @@ const IDX_HEADER_SIZE = 259 * 4 // 3 state uint32 + 256 chanCounts uint32
 
 export interface IdxPack {
    chanIds: Uint8Array
-   fetchedAts: Uint32Array
+   fetchedAts: Uint16Array
    chanCounts: Uint32Array
    ownChanCounts: Uint32Array
    bounds: { packId: number; startChron: number }[]
@@ -42,7 +42,7 @@ export function makeIdxPack(buf: ArrayBuffer, packIndex: number, packSize: numbe
       chanIds: new Uint8Array(0),
       chanCounts: new Uint32Array(0),
       ownChanCounts: new Uint32Array(0),
-      fetchedAts: new Uint32Array(0),
+      fetchedAts: new Uint16Array(0),
       bounds: [],
       parse() {
          if (!rawBuf) return pack
@@ -64,10 +64,10 @@ export function makeIdxPack(buf: ArrayBuffer, packIndex: number, packSize: numbe
          }
 
          const chanIds = new Uint8Array(packSize)
-         // Uint32 (not Uint16): fetchedAt accumulates 8h-blocks since
-         // first_fetched; a Uint16 ceiling wraps after ~60y of continuous
-         // fetching (65535 blocks), silently corrupting time-range jumps.
-         const fetchedAts = new Uint32Array(packSize)
+         // fetchedAt is 8h-blocks since first_fetched. Uint16 caps at 65535
+         // blocks ≈ 60y of calendar time from the first fetch — far past any
+         // real horizon, so the wrap it would eventually cause is acceptable.
+         const fetchedAts = new Uint16Array(packSize)
          pack.chanIds = chanIds
          pack.fetchedAts = fetchedAts
          const bytes = new Uint8Array(rawBuf)
